Add files via upload
This is a C++ program I worked on as a homework assignment. I was given the array2d.hpp file and had to write the rest of the code. The Makefile was also provided by my instructor. If you already have Linux installed on your machine, you can run array2d right away. 

This program generates a user defined number of rows and columns and for each row will produce the multiples of that row number (0-indexed).